+++ preview/template.ejs
@@ -40,5 +40,7 @@
   <iframe class="js-text preview-iframe preview-iframe--text preview-iframe--<%- template.TemplateType.toLowerCase() %> hidden" frameborder="0" src="/text/<%- template.Alias %>"></iframe>
 
   <script src="js/preview.js" type="text/javascript"></script>
+
+  <%- include("partials/hotReload.ejs") %>
 </body>
 </html>
